Separate the implementaion of MFCC and MFE
